requirements.txt
torch --index-url https://download.pytorch.org/whl/cpu
pygame-ce==2.5.4
python-chess